requirements.txt
asgiref==3.5.0
attrs==22.2.0
awsebcli==3.20.3
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blessed==1.20.0
boto3==1.20.54
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.0.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-environ==0.9.0
django-import-export==2.7.1
django-rest-passwordreset==1.3.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
ez-setup==0.9
filelock==3.8.0
future==0.16.0
gunicorn==20.1.0
identify==2.5.9
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
mailersend==0.5.0
MarkupPy==1.14
MarkupSafe==2.1.0
nodeenv==1.7.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
paramiko==3.0.0
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.4
pre-commit==2.20.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
qrcode==7.3.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
tablib==3.2.0
termcolor==1.1.0
texttable==1.6.7
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.17.0
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
